[add] statement and file_input
