+++ src/parser.rs
@@ -8,4 +8,6 @@ pub mod assign;
 pub mod call;
 pub mod index;
 pub mod member;
+pub mod simplestatement;
 pub mod statement;
+pub mod file_input;
